+++ src/fastmath/interpolation/monotone.clj
@@ -52,3 +52,4 @@
                          (* (- 3.0 t t) (Array/aget ys (inc i))))))))))))
 
 (m/unuse-primitive-operators)
+